requirements.txt
selenium == 4.38.0
pytest == 9.0.1
PyYAML == 6.0.3
python-dotenv == 1.2.1
Faker == 38.2.0
requests == 2.32.5
schema == 0.7.5
pytest-html == 4.1.1
pytest-rerunfailures == 16.1